pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cppgolf"
version = "0.1.8"
description = "C++ multi-file merge & code golf / minifier tool"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["cpp", "c++", "golf", "minify", "code-golf", "competitive-programming"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Text Processing :: Filters",
]

dependencies = [
    "libclang",
]

[project.optional-dependencies]
dev = [
    "build",
    "twine",
    "pytest",
]

[project.scripts]
cppgolf = "cppgolf.__main__:main"

[project.urls]
Homepage = "https://github.com/KnCRJVirX/cppgolf"
Issues   = "https://github.com/KnCRJVirX/cppgolf/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cppgolf*"]
